+++ src/main/java/com/incubyte/salary/dto/SalaryResponse.java
@@ -1,13 +1,14 @@
 package com.incubyte.salary.dto;
 
 import lombok.*;
+import java.math.BigDecimal; // Import for precision
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 public class SalaryResponse {
-    private Double grossSalary;
-    private Double deduction;
-    private Double netSalary;
+    private BigDecimal grossSalary;
+    private BigDecimal deduction;
+    private BigDecimal netSalary;
 }
